test(datetime): migrate root test
